add:type에 rank.ts추가
