+++ src/app/learning/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-const page = () => {
-  return <div>page</div>;
+const LearningPage = () => {
+  return <div>LearningPage</div>;
 };
 
-export default page;
+export default LearningPage;
