+++ src/types/result.ts
@@ -15,31 +15,6 @@ export interface GamesArray {
   name: string;
 }
 
-export interface RankIncludingUserInfo {
-  user_id: string;
-  checking: number;
-  speaking: number;
-  writing: number;
-  created_at: string;
-  id: string;
-  total: number;
-  week: number;
-  ranking: number;
-  user: { image: string; nickname: string; introduction: string };
-}
-
-export interface Rank {
-  user_id: string;
-  checking: number;
-  speaking: number;
-  writing: number;
-  created_at: string;
-  id: string;
-  total: number;
-  week: number;
-  ranking: number;
-}
-
 export interface matchedGameArrayForGuest {
   type: string;
   score: string | null;
